feat: Implement age-based ticket pricing and user-type-specific UI
This commit introduces two major features:
1. Age-based ticket pricing: The ticket booking system now adjusts prices based on the age of the customer, offering discounts for certain age groups.
2. User-type-specific UI: The interface now adapts based on the user type (e.g., admin, customer), providing a customized experience for each category.

These enhancements aim to improve user satisfaction and system flexibility.

+++ src/main/java/com/aeroBlasters/flightManagementSystem/bean/Flight.java
@@ -10,7 +10,7 @@ public class Flight {
 	private Long flightNumber;
 	private String carrierName;
 	private Long routeId;
-	private Integer seatCapacity;
+	private Long seatCapacity;
 	private String departure;
 	private String arrival;
 	private Long seatBooked;
@@ -43,11 +43,11 @@ public class Flight {
 		this.routeId = routeId;
 	}
 
-	public Integer getSeatCapacity() {
+	public Long getSeatCapacity() {
 		return seatCapacity;
 	}
 
-	public void setSeatCapacity(Integer seatCapacity) {
+	public void setSeatCapacity(Long seatCapacity) {
 		this.seatCapacity = seatCapacity;
 	}
 
@@ -75,7 +75,7 @@ public class Flight {
 		this.seatBooked = seatBooked;
 	}
 
-	public Flight(Long flightNumber, String carrierName, Long routeId, Integer seatCapacity, String departure,
+	public Flight(Long flightNumber, String carrierName, Long routeId, Long seatCapacity, String departure,
 			String arrival) {
 		super();
 		this.flightNumber = flightNumber;
